requirements.txt
cffi==1.14.6
click==8.0.1
colorclass==2.2.0
cryptography==3.4.7
easygui==0.98.2
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
msoffcrypto-tool==4.12.0
olefile==0.46
oletools==0.60
pcodedmp==1.2.6
pycparser==2.20
pyparsing==2.4.7
python-dotenv==0.19.0
python-magic==0.4.24
selenium==3.141.0
urllib3==1.26.6
Werkzeug==2.0.1
